+++ src/extensions.js
@@ -36,7 +36,7 @@ TableFrameMorph, console, Morph, MenuMorph*/
 
 /*jshint esversion: 11, bitwise: false*/
 
-modules.extensions = '2026-February-27';
+modules.extensions = '2026-March-02';
 
 // Global stuff
 
@@ -1416,6 +1416,12 @@ SnapExtensions.primitives.set(
     'cst_morph(cst)',
     function (costume, proc) {
         var m = new Morph(),
+            imageView = () => new DialogBoxMorph().inform(
+                'Image view',
+                null,
+                this.world(),
+                costume.contents
+            ),
             img;
         proc.assertType(costume, 'costume');
         img = costume.contents;
@@ -1425,6 +1431,8 @@ SnapExtensions.primitives.set(
         m.cachedImage = img;
         m.isCustomSwatch = true;
 
+        m.mouseDoubleClick = imageView;
+
         // support exporting costumes directly from speech balloons:
         m.userMenu = function () {
             var menu = new MenuMorph(this),
@@ -1432,6 +1440,7 @@ SnapExtensions.primitives.set(
                     this.world().childThatIsA(IDE_Morph);
 
             if (ide.isAppMode) {return; }
+            menu.addItem('open in dialog...', imageView);
             menu.addItem(
                 'export',
                 () => ide.saveCanvasAs(img, costume.name || 'image')